backend/requirements.txt
numpy>=1.10
scipy
Image
nltk
Flask
Flask-Uploads
img2pdf
python-env
opencv-python
flask-socketio
flask-cors
google.cloud
google-cloud-vision
python-dotenv
stanfordcorenlp

